requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-countries==7.5.1
django-typed-models==0.13.0
drf-typed==0.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.6.0